Extract TrackHistoryWithMatomo into own file
