+++ projects/bp-gallery/src/main.tsx
@@ -6,7 +6,8 @@ import { BrowserRouter } from 'react-router-dom';
 import App from './components/App';
 import { root } from './helpers/app-helpers';
 import './i18n';
-import setupMatomo, { TrackHistoryWithMatomo } from './matomo';
+import setupMatomo from './matomo/matomo';
+import { TrackHistoryWithMatomo } from './matomo/TrackHistoryWithMatomo';
 import reportWebVitals from './reportWebVitals';
 
 const sentryDsn = import.meta.env.VITE_REACT_APP_SENTRY_DSN;
